pyproject.toml
[project]
name = "aethermind"
version = "0.1.0"
description = ""
authors = [
    {name = "wutingkang",email = "tingkan2@illinois.edu"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "langchain (>=0.3.22,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)"
]

[tool.poetry]
packages = [{include = "aethermind", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
